requirements.txt
torch>=2.1.0
transformers>=4.33.0
pytest>=7.4.0
evaluate>=0.4.0
coverage>=7.2.7
python-dotenv>=1.0.0
pytest-cov>=4.1.0
rouge-score>=0.1.2
optree>=0.13.0
streamlit>=1.32.0
huggingface-hub>=0.28.1
accelerate>=0.28.0
bitsandbytes>=0.41.0
sentencepiece>=0.2.0 